vyatta-bmc-sel: don't error is no IPMI found
Since at least one the configuration nodes have a default value,
this script may run even if IPMI isn't configured. Exiting with
an error may block valid configuration. So just warn the user
that we can't do anything with the configuration--In case they
believe that they have IPMI devices and need to investigate.

This was exposed recently by some changes to the configuration
infrastructure.

Bugfix: VRVDR-52657