vyatta-bmc-sel: ignore BMC SEL on unsupported platforms
Attempting to configure the BMC SEL on some IPMI implementations fails in
a very noisy way, creating a commit failure. We only really have tested
and support this configuration on one platform.  Warn the user if we
attempt to change the BMC SEL when we are unsure about what we can do.

Bugfix: VRVDR-53277