Do not exit 1 if failed to detect known platform
The CLI "set system bmc sel mode capacity" is run on system start-up
even if there is no bmc config, as it is a default value. On systems
with IPMI but not a supported hardware platform it causes the following
message in the log, and then exits with 1.

 failed to detect platform PlatformError('no platform detected')

When the configuration system gets a failed config (due to exit 1),
it then skips any other config at the same priority (which would be the
default priority of 0). This results in config being skipped, for this
particular ticket it skips the "system session" configuration.

The fix here is to exit 0, as it should be expected that the platform
not be detected on systems, and so allow the config to proceed.

DAN-462